api-gateway/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
google-auth>=2.0.0
requests>=2.32.0
neo4j>=5.14.0
pandas>=2.0.0
